% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyNA.R
\name{anyNA}
\alias{anyNA}
\title{Backport of anyNA}
\usage{
anyNA(x, recursive = FALSE)
}
\arguments{
\item{x}{See description.}

\item{recursive}{See description.}
}
\description{
See the original description in \code{base::anyNA}.
}

