% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rconcordance.R
\name{as.character.Rconcordance}
\alias{as.character.Rconcordance}
\title{Backport of as.character.Rconcordance for R < 4.3.0}
\usage{
\method{as.character}{Rconcordance}(x, targetfile = "", ...)
}
\description{
See the original description in \code{tools::as.character.Rconcordance}.
}
\keyword{internal}
