% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rconcordance.R
\name{as.Rconcordance.default}
\alias{as.Rconcordance.default}
\title{Backport of as.Rconcordance.default for R < 4.3.0}
\usage{
\method{as.Rconcordance}{default}(x, ...)
}
\description{
See the original description in \code{tools::as.Rconcordance}.
}
\keyword{internal}
