% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rconcordance.R
\name{as.Rconcordance}
\alias{as.Rconcordance}
\title{Backport of as.Rconcordance for R < 4.3.0}
\usage{
as.Rconcordance(x, ...)
}
\description{
See the original description in \code{tools::as.Rconcordance}.
}
\keyword{internal}
