% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asplit.R
\name{asplit}
\alias{asplit}
\title{Backport of asplit for R < 3.6.0}
\usage{
asplit(x, MARGIN)
}
\description{
See the original description in \code{base::asplit}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_asplit = getFromNamespace("asplit", "backports")
x = matrix(1:6, 2, 3)
bp_asplit(x, 1)
}
\keyword{internal}
