% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{backports-package}
\alias{backports}
\alias{backports-package}
\title{backports: Reimplementations of Functions Introduced Since R-3.0.0}
\description{

}
\section{Backports for R versions prior to 3.2.0}{

\itemize{
 \item{\code{\link{anyNA}}}
 \item{\code{\link{dir.exists}}}
 \item{\code{\link{file.size}}, \code{\link{file.mode}}, \code{\link{file.mtime}}}
 \item{\code{\link{lengths}}}
 \item{Argument} \code{extra_cols} for \code{\link{file.info}}
}
}

\section{Backports for R versions prior to 3.3.0}{

\itemize{
 \item{\code{\link{startsWith}} and \code{\link{endsWith}}}
 \item{\code{\link{strrep}}}
 \item{\code{\link{trimws}}}
 \item{Argument \code{type} for \code{\link{capture.output}}}
}
}

\section{Backports for R versions prior to 3.4.0}{

\itemize{
 \item{\code{\link{.valid.factor}}}
 \item{\code{\link{hasName}} from \pkg{utils}}
}
}

\section{Backports for R versions prior to 3.5.0}{

\itemize{
 \item{\code{\link{...length}}}
 \item{\code{\link{...elt}}}
 \item{\code{\link{isFALSE}}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/backports}
  \item Report bugs at \url{https://github.com/r-lib/backports/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (0000-0001-9754-0393)

Authors:
\itemize{
  \item R Core Team
}

}
