% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.output.R
\name{capture.output}
\alias{capture.output}
\title{Backport of capture.output for R < 3.3.0}
\usage{
capture.output(
  ...,
  file = NULL,
  append = FALSE,
  type = c("output", "message"),
  split = FALSE
)
}
\description{
Argument \code{type} has been backported.
See the original description in \code{utils::capture.output}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_capture.output <- getFromNamespace("capture.output", "backports")

captured <- bp_capture.output({ message("hi") }, type = "message")
str(captured)
}
\keyword{internal}
