% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotsElt.R
\name{...elt}
\alias{...elt}
\title{Backport of ...elt for R < 3.4.2}
\usage{
...elt(n)
}
\description{
See the original description in \code{base::...elt}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_...elt = getFromNamespace("...elt", "backports")

foo = function(n, ...) bp_...elt(n)
foo(n = 2, "a", "b", "c")
}
\keyword{internal}
