% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotsLength.R
\name{...length}
\alias{...length}
\title{Backport of ...length for R < 3.4.2}
\usage{
...length()
}
\description{
See the original description in \code{base::...length}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_...length = getFromNamespace("...length", "backports")

foo = function(...) bp_...length()
foo(1, 2, 3)
}
\keyword{internal}
