% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotsNames.R
\name{...names}
\alias{...names}
\title{Backport of ...names for R < 4.1.0}
\usage{
...names()
}
\description{
See the original description in \code{base::...names}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_...names = getFromNamespace("...names", "backports")

foo = function(...) bp_...names()
foo(a = 1, b = 2, 3)
}
\keyword{internal}
