% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dQuote.R
\name{dQuote}
\alias{dQuote}
\alias{sQuote}
\title{Backport of dQuote and sQuote for R < 3.6.0}
\usage{
dQuote(x, q = getOption("useFancyQuotes"))

sQuote(x, q = getOption("useFancyQuotes"))
}
\description{
Argument \code{q} has been backported.
See the original description in \code{base::dQuote()} and \code{base::sQuote()}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_dQuote <- getFromNamespace("dQuote", "backports")
bp_dQuote("foo")
bp_dQuote("foo", q = TRUE)
}
\keyword{internal}
