% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endsWith.r
\name{endsWith}
\alias{endsWith}
\title{Backport of endsWith}
\usage{
endsWith(x, suffix)
}
\arguments{
\item{x}{See description.}

\item{suffix}{See description.}
}
\description{
See the original description in \code{base::endsWith}.
}

