% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.info.R
\name{file.info}
\alias{file.info}
\title{Backport of \code{file.info}}
\usage{
file.info(..., extra_cols = TRUE)
}
\arguments{
\item{...}{See description.}

\item{extra_cols}{See description.}
}
\description{
Argument \code{extra_cols} has been backported.
See the original description in \code{base::file.info}.
}

