% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.x.r
\name{file.size}
\alias{file.mode}
\alias{file.mtime}
\alias{file.size}
\title{Backports of wrappers around \code{file.info}}
\usage{
file.size(...)

file.mode(...)

file.mtime(...)
}
\arguments{
\item{...}{See description.}
}
\description{
See the original description in \code{base::file.size}.
}

