% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rconcordance.R
\name{followConcordance}
\alias{followConcordance}
\title{Backport of followConcordance for R < 4.3.0}
\usage{
followConcordance(concordance, prevConcordance)
}
\description{
See the original description in \code{tools::followConcordance}.
}
\keyword{internal}
