% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get0.R
\name{get0}
\alias{get0}
\title{Backport of get0 for R < 3.2.0}
\usage{
get0(
  x,
  envir = pos.to.env(-1L),
  mode = "any",
  inherits = TRUE,
  ifnotfound = NULL
)
}
\description{
See the original description in \code{base::get0}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_get0 = getFromNamespace("get0", "backports")

bp_get0("a")
bp_get0("a", ifnotfound = 0)

foo = 12
bp_get0("foo")
}
\keyword{internal}
