% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasName.R
\name{hasName}
\alias{hasName}
\title{Backport of hasName}
\usage{
hasName(x, name)
}
\arguments{
\item{x}{See description.}

\item{name}{See description.}
}
\description{
See the original description in \code{utils::hasName}.
}

