% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFALSE.R
\name{isFALSE}
\alias{isFALSE}
\title{Backport of isFALSE for R < 3.4.2}
\usage{
isFALSE(x)
}
\description{
See the original description in \code{base::isFALSE}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_isFALSE = getFromNamespace("isFALSE", "backports")

bp_isFALSE(FALSE)
bp_isFALSE(iris)
}
\keyword{internal}
