% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNamespaceLoaded.R
\name{isNamespaceLoaded}
\alias{isNamespaceLoaded}
\title{Backport of isNamespaceLoaded for R < 3.2.0}
\usage{
isNamespaceLoaded(name)
}
\description{
See the original description in \code{base::isNamespaceLoaded}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_isNamespaceLoaded = getFromNamespace("isNamespaceLoaded", "backports")

bp_isNamespaceLoaded("backports")
}
\keyword{internal}
