% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isS3stdGeneric.R
\name{isS3stdGeneric}
\alias{isS3stdGeneric}
\title{Backport of isS3stdGeneric for R < 4.1.0}
\usage{
isS3stdGeneric(f)
}
\description{
See the original description in \code{utils::isS3stdGeneric}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_isS3stdGeneric <- getFromNamespace("isS3stdGeneric", "backports")

bp_isS3stdGeneric(mean)

f <- function(x) x
bp_isS3stdGeneric(f)
}
\keyword{internal}
