% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengths.R
\name{lengths}
\alias{lengths}
\title{Backport of lengths for R < 3.2.0}
\usage{
lengths(x, use.names = TRUE)
}
\description{
See the original description in \code{base::lengths}.
}
\examples{
bp_lengths = getFromNamespace("lengths", "backports")
bp_lengths(list(1:3, 2))
}
\keyword{internal}

