% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libPaths.R
\name{.libPaths}
\alias{.libPaths}
\title{Backport of .libPaths for R < 4.1.0}
\usage{
.libPaths(new, include.site = TRUE)
}
\description{
See the original description in \code{base::.libPaths}.
}
\examples{
save <- .libPaths()
save
# ignore the site library
.libPaths("test", include.site = FALSE)

# restore the original
.libPaths(save)
}
\keyword{internal}
