% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2DF.R
\name{list2DF}
\alias{list2DF}
\title{Backport of list2DF for R < 4.0.0}
\usage{
list2DF(x = list(), nrow = NULL)
}
\description{
See the original description in \code{base::R_user_dir}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_list2DF = getFromNamespace("list2DF", "backports")

## Create a data frame holding a list of character vectors and the
## corresponding lengths:
x <- list(character(), "A", c("B", "C"))
n <- lengths(x)
bp_list2DF(list(x = x, n = n))

## Create data frames with no variables and the desired number of rows:
bp_list2DF()
bp_list2DF(nrow = 3L)
}
\keyword{internal}
