% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rconcordance.R
\name{matchConcordance}
\alias{matchConcordance}
\title{Backport of matchConcordance for R < 4.3.0}
\usage{
matchConcordance(linenum, concordance)
}
\description{
See the original description in \code{tools::matchConcordance}.
}
\keyword{internal}
