% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{notin_operator}
\alias{notin_operator}
\alias{\%notin\%}
\title{Backport of the not-in operator for R < 4.6.0}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{(any).}

\item{table}{(any).}
}
\description{
See the original description in \code{match}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_notin = getFromNamespace("\%notin\%", "backports")

bp_notin(1, 2:4)
}
\keyword{internal}
