% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_coalesce.R
\name{null_coalesce_operator}
\alias{null_coalesce_operator}
\alias{\%||\%}
\title{Backport of the null-coalescing operator for R < 4.4.0}
\usage{
x \%||\% y
}
\arguments{
\item{x}{(any).}

\item{y}{(any).}
}
\description{
See the original description in \code{Control}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_null_coalesce = getFromNamespace("\%||\%", "backports")

bp_null_coalesce(NULL, FALSE)
}
\keyword{internal}
