% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullfile.R
\name{nullfile}
\alias{nullfile}
\title{Backport of nullfile for R < 3.6.0}
\usage{
nullfile()
}
\description{
See the original description in \code{base::nullfile}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_nullfile = getFromNamespace("nullfile", "backports")

bp_nullfile()
}
\keyword{internal}
