% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{order}
\alias{order}
\title{Backport of \code{order} for R < 3.3.0}
\usage{
order(..., na.last = TRUE, decreasing = FALSE, method = c("shell",
  "radix"))
}
\description{
Allows value \code{"radix"} for argument \code{method}.
See the original description in \code{base::order}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_order = getFromNamespace("order", "backports")

x = runif(10)
data.frame(x = x, order = bp_order(x, method = "radix"))
}
\keyword{internal}

