% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.R
\name{paste}
\alias{paste}
\title{Backport of paste for R < 4.0.1}
\usage{
paste(..., sep = " ", collapse = NULL, recycle0 = FALSE)
}
\description{
See the original description in \code{base::paste}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_paste = getFromNamespace("paste", "backports")
bp_paste(letters[1:3], character(), collapse = NULL, recycle0 = TRUE)
}
\keyword{internal}
