% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste0.R
\name{paste0}
\alias{paste0}
\title{Backport of paste0 for R < 4.0.1}
\usage{
paste0(..., collapse = NULL, recycle0 = FALSE)
}
\description{
See the original description in \code{base::paste0}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_paste0 = getFromNamespace("paste0", "backports")
bp_paste0(letters[1:3], character(), collapse = NULL, recycle0 = TRUE)
}
\keyword{internal}
