% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rconcordance.R
\name{print.Rconcordance}
\alias{print.Rconcordance}
\title{Backport of print.Rconcordance for R < 4.3.0}
\usage{
\method{print}{Rconcordance}(x, ...)
}
\description{
See the original description in \code{tools::print.Rconcordance}.
}
\keyword{internal}
