% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_user_dir.R
\name{R_user_dir}
\alias{R_user_dir}
\title{Backport of R_user_dir for R < 4.0.0}
\usage{
R_user_dir(package, which = c("data", "config", "cache"))
}
\description{
See the original description in \code{tools::R_user_dir}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_R_user_dir = getFromNamespace("R_user_dir", "backports")

bp_R_user_dir("backports")
}
\keyword{internal}
