% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSource.R
\name{removeSource}
\alias{removeSource}
\title{Backport of removeSource for R < 3.6.0}
\usage{
removeSource(fn)
}
\description{
See the original description in \code{utils::removeSource}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_removeSource = getFromNamespace("removeSource", "backports")

bp_removeSource(mean)
}
\keyword{internal}
