% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort}
\alias{sort}
\title{Backport of \code{sort} for R < 3.3.0}
\usage{
sort(x, decreasing = FALSE, method = c("shell", "quick", "radix"), ...)
}
\description{
Allows value \code{"radix"} for argument \code{method}.
See the original description in \code{base::sort}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_sort = getFromNamespace("sort", "backports")

bp_sort(runif(10), method = "radix")
}
\keyword{internal}

