% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startsWith.r
\name{startsWith}
\alias{startsWith}
\title{Backport of startsWith}
\usage{
startsWith(x, prefix)
}
\arguments{
\item{x}{See description.}

\item{prefix}{See description.}
}
\description{
See the original description in \code{base::startsWith}.
}

