% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopifnot.R
\name{stopifnot}
\alias{stopifnot}
\title{Backport of stopifnot for R versions < 4.0.0.}
\usage{
stopifnot(..., exprs, exprObject, local = TRUE)
}
\description{
See the original description in \code{base::stopifnot}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_stopifnot = getFromNamespace("stopifnot", "backports")

m <- matrix(c(1, 3, 12, 1), 2, 2)
\dontrun{bp_stopifnot("m must be symmetric" = m == t(m))}
}
\keyword{internal}
