% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str2expression.R
\name{str2expression}
\alias{str2expression}
\title{Backport of str2expression for R < 3.6.0}
\usage{
str2expression(text)
}
\description{
See the original description in \code{base::str2expression}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
str2expression <- getFromNamespace("str2expression", "backports")

str2expression("x[3] <- 1+4")
}
\keyword{internal}
