% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str2lang.R
\name{str2lang}
\alias{str2lang}
\title{Backport of str2lang for R < 3.6.0}
\usage{
str2lang(s)
}
\description{
See the original description in \code{base::str2lang}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
str2lang <- getFromNamespace("str2lang", "backports")

str2lang("x[3] <- 1+4")
}
\keyword{internal}
