% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strcapture.R
\name{strcapture}
\alias{strcapture}
\title{Backport of strcapture st2DF for R < 4.0.0}
\usage{
strcapture(pattern, x, proto, perl = FALSE, useBytes = FALSE)
}
\description{
See the original description in \code{utils::strcapture}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_strcapture = getFromNamespace("strcapture", "backports")

bp_strcapture("backports")
}
\keyword{internal}
