% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strrep.R
\name{strrep}
\alias{strrep}
\title{Backport of strrep}
\usage{
strrep(x, times)
}
\arguments{
\item{x}{See description.}

\item{times}{See description.}
}
\description{
See the original description in \code{base::strrep}.
}

