% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppressWarnings.R
\name{suppressWarnings}
\alias{suppressWarnings}
\alias{suppressMessages}
\title{Backport of suppressWarnings and suppressMessages for R < 4.0.0}
\usage{
suppressWarnings(expr, classes = "warning")

suppressMessages(expr, classes = "message")
}
\description{
Backport for new argument `classes`.
See the original description in \code{base::suppressWarnings}/\code{base::suppressMessages}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_suppressWarnings = getFromNamespace("suppressWarnings", "backports")
bp_suppressWarnings(warningCondition("warning", class = "testWarning"), "testWarning")
}
\keyword{internal}
