% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimws.R
\name{trimws}
\alias{trimws}
\title{Backport of trimws}
\usage{
trimws(x, which = c("both", "left", "right"))
}
\arguments{
\item{x}{See description.}

\item{which}{See description.}
}
\description{
See the original description in \code{base::trimws}.
}

