% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URLencode.R
\name{URLencode}
\alias{URLencode}
\title{Backport of URLencode for R < 3.2.0}
\usage{
URLencode(URL, reserved = FALSE, repeated = FALSE)
}
\description{
See the original description in \code{utils::URLencode}.
Argument \code{repeated} is backported.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_URLencode = getFromNamespace("URLencode", "backports")

URLdecode(z <- "ab\%20cd")
c(bp_URLencode(z), bp_URLencode(z, repeated = TRUE))
}
\keyword{internal}
