% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.factor.R
\name{.valid.factor}
\alias{.valid.factor}
\title{Backport of .valid.factor for R < 3.4.0}
\usage{
.valid.factor(object)
}
\description{
See the original description in \code{base::.valid.factor}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_.valid_factor = getFromNamespace(".valid.factor", "backports")
bp_.valid_factor(factor(letters[1:3]))
}
\keyword{internal}
