% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignetteInfo.R
\name{vignetteInfo}
\alias{vignetteInfo}
\title{Backport of vignetteInfo for R < 3.6.0}
\usage{
vignetteInfo(file)
}
\description{
See the original description in \code{tools::vignetteInfo}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_vignetteInfo = getFromNamespace("vignetteInfo", "backports")
}
\keyword{internal}
