% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyNA.R
\name{anyNA}
\alias{anyNA}
\title{Backport of anyNA for R versions < 3.2.0.}
\usage{
anyNA(x, recursive = FALSE)
}
\description{
See the original description in \code{base::anyNA}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_anyNA = getFromNamespace("anyNA", "backports")

bp_anyNA(letters)
}
\keyword{internal}
