% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{package}
\name{backports-package}
\alias{backports}
\alias{backports-package}
\title{This package provides functions which have been introduced in one of the base packages in R version 3.0.1 or later.
These backports are conditionally exported in order to let R resolve the function name to the implemented backport only if the respective base version is not available.
As a package developer, you can just import \code{backports} to use all implemented backports without worrying about the required R version.}
\description{
This package provides functions which have been introduced in one of the base packages in R version 3.0.1 or later.
These backports are conditionally exported in order to let R resolve the function name to the implemented backport only if the respective base version is not available.
As a package developer, you can just import \code{backports} to use all implemented backports without worrying about the required R version.
}

