% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{package}
\name{backports-package}
\alias{backports}
\alias{backports-package}
\title{Backports of Functions Introduced Since R-3.0.0}
\description{
This package provides backports of functions which have been introduced in one of the base packages in R version 3.0.1 or later.
These backports are conditionally exported in order to let R resolve the function name to either the implemented backport, or the respective base version, if available.
This way, package developers can make use of the new, optimized functions without without worrying about the minimum R version required.
}

