% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{backports-package}
\alias{backports}
\alias{backports-package}
\title{backports: Reimplementations of Functions Introduced Since R-3.0.0}
\description{
Functions introduced or changed since R v3.0.0 are re-implemented in this package. The backports are conditionally exported in order to let R resolve the function name to either the implemented backport, or the respective base version, if available. Package developers can make use of new functions or arguments by selectively importing specific backports to support older installations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/backports}
  \item Report bugs at \url{https://github.com/r-lib/backports/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})

Authors:
\itemize{
  \item R Core Team
}

}
