% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.exists.r
\name{dir.exists}
\alias{dir.exists}
\title{Backport of lengths}
\usage{
dir.exists(paths)
}
\arguments{
\item{paths}{See description.}
}
\description{
See the original description in \code{base::dir.exists}.
}

