% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endsWith.R
\name{endsWith}
\alias{endsWith}
\title{Backport of endsWith for R < 3.3.0}
\usage{
endsWith(x, suffix)
}
\description{
See the original description in \code{base::endsWith}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_endsWith = getFromNamespace("endsWith", "backports")

bp_endsWith(c("aabb", "bbcc"), "bb")
}
\keyword{internal}
