% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.info.R
\name{file.info}
\alias{file.info}
\title{Backport of \code{file.info} for R < 3.2.0}
\usage{
file.info(..., extra_cols = TRUE)
}
\description{
Argument \code{extra_cols} has been backported.
See the original description in \code{base::file.info}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_file.info = getFromNamespace("file.info", "backports")

bp_file.info(file.path(R.home(), "COPYING"), extra_cols = FALSE)
}
\keyword{internal}
