% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.mode.R, R/file.mtime.R, R/file.size.R
\name{file.mode}
\alias{file.mode}
\alias{file.mtime}
\alias{file.size}
\title{Backports of wrappers around \code{file.info}}
\usage{
file.mode(...)

file.mtime(...)

file.size(...)
}
\arguments{
\item{...}{See description.}
}
\description{
See the original description in \code{base::file.size}.
}

