% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasName.R
\name{hasName}
\alias{hasName}
\title{Backport of hasName for R < 3.4.0}
\usage{
hasName(x, name)
}
\description{
See the original description in \code{utils::hasName}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_hasName = getFromNamespace("hasName", "backports")

bp_hasName(list(a = 1, b = 2), c("a", "b", "c"))
}
\keyword{internal}
