% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengths.R
\name{lengths}
\alias{lengths}
\title{Backport of lengths}
\usage{
lengths(x, use.names = TRUE)
}
\arguments{
\item{x}{See description.}

\item{use.names}{See description.}
}
\description{
See the original description in \code{base::lengths}.
}

