% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2DF.R
\name{list2DF}
\alias{list2DF}
\title{Backport of list2DF for R < 4.0.0}
\usage{
list2DF(x = list(), nrow = NULL)
}
\description{
See the original description in \code{base::list2DF}.
}
\examples{
# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_list2DF = getFromNamespace("list2DF", "backports")

bp_list2DF(list(1:3, 2))
}
\keyword{internal}
